Replace src with testsrc, fix filtering logic error
