+++ frontend/src/App.tsx
@@ -1,27 +1,32 @@
-import { CssBaseline, Box, ThemeProvider } from '@mui/material';
-import MainPage from './components/MainPage.tsx';
-import theme from './theme/theme';
+import {type FC} from "react";
+import {Box, CssBaseline, ThemeProvider} from "@mui/material";
+import theme from "./theme/theme.ts";
+import MainPage from "./MainPage";
+import {Flex} from "./components/FlexGrid.tsx";
+import {QueryClient, QueryClientProvider} from "@tanstack/react-query";
 
-function App() {
+const client = new QueryClient()
+const App: FC = () => {
     return (
-        <ThemeProvider theme={theme}>
-            <CssBaseline />
-            <Box
-                sx={{
-                    display: 'flex',
-                    flexDirection: 'column',
-                    minHeight: '100vh',
-                    width: '100%',
-                    maxWidth: '100vw',
-                    overflowX: 'hidden'
-                }}
-            >
-                <Box sx={{ flex: '1 0 auto', py: { xs: 2, sm: 3, md: 4 }, width: '100%' }}>
-                    <MainPage />
-                </Box>
-            </Box>
-        </ThemeProvider>
-    );
+        <QueryClientProvider client={client}>
+            <ThemeProvider theme={theme}>
+                <CssBaseline/>
+                <Flex
+                    sx={{
+                        flexDirection: 'column',
+                        minHeight: '100vh',
+                        width: '100%',
+                        maxWidth: '100vw',
+                        overflowX: 'hidden'
+                    }}
+                >
+                    <Box sx={{flex: '1 0 auto', py: {xs: 2, sm: 3, md: 4}, width: '100%'}}>
+                        <MainPage/>
+                    </Box>
+                </Flex>
+            </ThemeProvider>
+        </QueryClientProvider>
+    )
 }
 
 export default App;
