+++ frontend/src/MainPage/CoursesArea/Course.tsx
