+++ frontend/src/MainPage/CoursesArea/Label.tsx
