+++ frontend/src/MainPage/CoursesArea/used/Description.tsx
