+++ frontend/src/MainPage/CoursesArea/used/DifficultyLabel.tsx
