+++ frontend/src/MainPage/CoursesArea/used/SubjectLabel.tsx
