+++ frontend/src/MainPage/FilterCard/used/Difficulties.tsx
