+++ frontend/src/components/DynamicSvg.tsx
