+++ frontend/src/main.tsx
@@ -1,7 +1,8 @@
 import {StrictMode} from 'react'
 import {createRoot} from 'react-dom/client'
-import './index.css'
 import App from './App.tsx'
+
+import './index.css'
 import {HashRouter} from "react-router-dom";
 
 const root = createRoot(
