+++ frontend/src/types/course.ts
