+++ frontend/src/types/filters/difficulty.ts
